refactor(OpenGL): moving rendering functions from Window to Render
